2023 Day 11 Perl O(n)
Based on observations:

 * X and Y can be calculated independently, and

 * rather than getting bool for empty rows/columns count #s and
   total #s

Consider counts of x values, the first column with non-zero
must step to the next column with #s <total count #s> minus
<number of hashes in first column> times. Add the first column
count to the second column count and repeat treating the second
column like the first. If the count of columns is zero we increase
the expansion count; resetting it when we "move" entries.